v15.2 Домашнее задание спринт №15

+++ app/src/main/java/com/practicum/playlistmaker/searchMusic/domain/entities/Track.kt
@@ -5,7 +5,7 @@ import java.util.Locale
 
 data class Track(val trackName: String,
                  val artistName: String,
-                 val trackTimeMillis: Long,
+                 val trackTimeMillis: Int,
                  val artworkUrl100: String,
                  val trackId: Int,
                  val collectionName: String,
